+++ Engine/Engine/MathFunctions.h
@@ -1,7 +1,7 @@
 #ifndef MATHFUNCTIONS_H
 #define MATHFUNCTIONS_H
 
-bool valueInRange(int value, int min, int max);
+bool valueInRange(float value, float min, float max);
 
 
 #endif
